Compute overtakes during data fetching
